+++ CHANGELOG.md
@@ -8,3 +8,4 @@ This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Implement checking for duplicates, big files, and placeholder images
 - Add console and SARIF output
 - Naively create `asset_lint_list.json` file
+- Total reclaimable size printed to console